requirements.txt
# requirements.txt

openai
pandas
numpy
scipy
scikit-learn
python-dotenv
